backend/pyproject.toml
[tool.poetry]
name = "macro-lens-backend"
version = "0.1.0"
description = "Macro Lens backend"
authors = ["You"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
httpx = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pandas = "^2.2.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
sendgrid = "^6.11.0"
apscheduler = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
